빠른 병합 하셔야합니다

+++ src/main/java/com/ohgiraffers/warehousemanagement/wms/auth/model/AuthDetails.java
@@ -101,7 +101,7 @@ public class AuthDetails implements UserDetails {
     }
 
     // id 반환
-    public Integer getUserId() {
+    public Long getUserId() {
         return loginUserDTO.getUserId();
     }
