+++ src/main/java/com/ohgiraffers/warehousemanagement/wms/purchases/model/entity/Purchase.java
@@ -32,7 +32,7 @@ public class Purchase {
 
     // 발주 담당자
     @Column(name = "user_id")
-    private Integer userId;
+    private Long userId;
 
     @Column(name = "purchase_date")
     private LocalDate purchaseDate;
@@ -63,10 +63,10 @@ public class Purchase {
 
 
     public Purchase(
-                    Integer purchaseId,
-                    Integer userId, LocalDate purchaseDate, LocalDate purchaseDueDate,
-                    PurchaseStatus purchaseStatus, LocalDate purchaseCreatedAt,
-                    LocalDate purchase_updated_at, String notes) {
+            Integer purchaseId,
+            Long userId, LocalDate purchaseDate, LocalDate purchaseDueDate,
+            PurchaseStatus purchaseStatus, LocalDate purchaseCreatedAt,
+            LocalDate purchase_updated_at, String notes) {
         //Supplier supplier, Storages storages,
 
 
@@ -125,11 +125,11 @@ public class Purchase {
         this.purchaseId = purchaseId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
